refactor: 프로젝트 설정 및 프로젝트 생성 UI div태그 수정
